[Android] HelpAndFeedback was removed
Chromium change:
https://chromium.googlesource.com/chromium/src/+/8098bca24a5d3610403552256d0e8bc4c9610429

Remove HelpAndFeedback.

The plan for this two-sided patch is:
1) Add the new class under feedback/ and AppHooks method upstream with
no reference.
2) Extend the new class downstream and add the new AppHooks method.
3) Update upstream references to point to the new AppHooks method.
4) Cleanup in downstream repository.
5) Cleanup in upstream repository.

1) was done in https://crrev.com/c/2364581.
2) was done in https://crrev.com/i/3227761.
3) was done in https://crrev.com/c/2414779.
4) was done in https://crrev.com/i/3274398.
5) is done in this CL.

Note that the change in SigninUtils was a miss from step 3). Ideally
all references to HelpAndFeedback should be removed in 3), but there is
one left in SigninUtils. The consequence is that after step 4) is landed,
this reference can only be pointed to the upstream implementation.

Bug: 1117343